requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
idna==2.9
requests==2.23.0
urllib3==1.25.8
click==7.1.1
fastapi==0.54.0
h11==0.9.0
httptools==0.1.1
psycopg2-binary==2.8.4
pydantic==1.4
python-dotenv==0.12.0
SQLAlchemy==1.3.13
starlette==0.13.2
uvicorn==0.11.3
uvloop==0.14.0
websockets==8.1
